// src/types.ts
/**
 * @file 定义类型和接口对象
 * @author zhaoyadong
 */

/**
 * 文本选项对象
 */
export interface ITextOption {
    // 年份文本描述
    yearText?: string;
    // 月份文本描述
    monthText?: string;
    // 日文本描述
    dayText?: string;
    // 小时文本描述
    hourText?: string;
    // 分钟文本描述
    minText?: string;
    // 秒钟文本描述
    secondText?: string;
};